MVTT-executable-app/requirements.txt
Pillow>=10.0.0
screeninfo>=0.8.1
pyinstaller>=6.0.0
